refactor automatic algorithms
