+++ swan/automatic_mapping.py
@@ -613,3 +613,349 @@ class ParameterInputDialog(QtWidgets.QDialog):
         if self._solvers is not None:
             self.chosen_solver = self._solvers[self._clusters.index(self.current_clusters_value)]
         return self.values_dict, self.chosen_solver, self.final_state
+
+
+def calculate_mapping_bu(virtual_unit_map, data, storage):
+    """
+    Calculates a mapping for the units based on features like distance.
+
+    The units will be compared pare-wise and sequential.
+
+    **Arguments**
+
+        *data* (:class:`src.neodata.NeoData`):
+            This object is needed to get the data
+            which will be used to compare the units.
+
+    """
+    wave_length = data.get_wave_length()
+
+    for i in range(len(data.blocks) - 1):
+        sessions = np.zeros((sum(data.total_units_per_block), 2, wave_length))
+
+        for j, val in enumerate(storage.get_map().mapping[i]):
+            if val != 0:
+                runit = virtual_unit_map.get_realunit(i, j, data)
+                avg = data.get_data("average", runit)
+                sessions[j][0] = avg / np.max(avg)
+            else:
+                sessions[j][0] = np.zeros(wave_length)
+
+        for j, val in enumerate(storage.get_map().mapping[i + 1]):
+            if val != 0:
+                runit = virtual_unit_map.get_realunit(i + 1, j, data)
+                avg = data.get_data("average", runit)
+                sessions[j][1] = avg / np.max(avg)
+            else:
+                sessions[j][1] = np.zeros(wave_length)
+
+        distances = cdist(sessions[:, 0], sessions[:, 1], metric='euclidean')
+        threshold = np.mean(distances[distances > 0])
+        print(threshold)
+        extend = 0
+        exclude = []
+        for j, val in enumerate(storage.get_map().mapping[i + 1]):
+            print("Executing this in session {}".format(i))
+            print("J: {}, Val: {}".format(j, val))
+
+            if val != 0:
+                print(distances[j])
+
+                min_arg = np.argmin(distances[j])
+                print("Min Arg: {}".format(min_arg))
+
+                if min_arg == j:
+                    pass
+                elif distances[j, min_arg] < threshold and (j, min_arg) not in exclude:
+                    print("Swapping {} and {}".format(j, min_arg))
+                    virtual_unit_map.swap(i + 1, j, min_arg)
+                    exclude.append((j, min_arg))
+                    exclude.append((min_arg, j))
+                elif distances[j, min_arg] >= threshold:
+                    print("Swapping {} and {}".format(j, data.total_units_per_block[i + 1] + extend))
+                    virtual_unit_map.swap(i + 1, j, data.total_units_per_block[i + 1] + extend)
+                    extend += 1
+                    print(extend)
+                else:
+                    print("Logic flawed, check again!")
+        print(exclude)
+    storage.change_map()
+
+
+def swan_implementation(virtual_unit_map, data, storage):
+    swaps = 0
+    # Retrieve mapping from base
+    backup_mapping = np.array(storage.get_map().mapping.copy()).T
+    pre_mapping = backup_mapping.copy()
+    # loop over columns/sessions
+    for s in range(pre_mapping.shape[1]):
+        # initialize array of mean waveform max amplitudes
+        averages = np.zeros((pre_mapping.shape[0]))
+
+        # loop over rows/units
+        for j in range(len(averages)):
+            # ignore zeros in the map (correspond to non-existent units)
+            if pre_mapping[j][s] != 0:
+                # Retrieve unit
+                runit = virtual_unit_map.get_realunit(s, j, data)
+                # Retrieve mean waveform for the unit
+                wave = data.get_data("average", runit)
+                # Store max amplitude of mean waveform
+                averages[j] = np.amax(np.abs(wave))
+
+        #
+        # Sort the units in each session by max amplitude of mean waveform
+        # and swap the corresponding units in the base.
+        #
+
+        # loop over rows/units
+        for i in range(len(averages)):
+            # retrieve index of max value in session
+            max_index = np.argmax(averages[i:]) + i
+            if max_index != i:
+                virtual_unit_map.swap(s, i, max_index)
+                tmp = averages[max_index]
+                averages[max_index] = averages[i]
+                averages[i] = tmp
+                swaps += 1
+
+    storage.change_map()
+    # Refresh the mapping using the values in base
+    mapping = np.array(storage.get_map().mapping.copy()).T
+
+    # Initialize an array shaped like "mapping" and set all values to np.nan
+    history = np.zeros_like(mapping, dtype=np.float)
+    history[history == 0.0] = None
+    print("Initial Mapping: {}".format(mapping))
+
+    # Loop over all columns/sessions except the last one
+    for i in range(mapping.shape[1] - 1):
+        # Loop over all rows/units
+        print("\nFirst loop: {}".format(i))
+        for j in range(mapping.shape[0]):
+            # ignore zeros in the map (correspond to non-existent units)
+            print("\tSecond Loop: {}".format(j))
+            if mapping[j][i] > 0:  # to avoid errors in float comparison
+                # Choose (j, i)th unit as the actor and obtain it's mean waveform
+                runit_actor = virtual_unit_map.get_realunit(i, j, data)
+                actor_wave = data.get_data("average", runit_actor)
+                actor_rp = data.get_data("rate profile", runit_actor)
+
+                # Initialize matrix to store distances to the actor unit
+                distances_waves = np.zeros((mapping.shape[0], mapping.shape[1]))
+                distances_rp = np.zeros((mapping.shape[0], mapping.shape[1]))
+
+                # Loop over all columns/sessions again
+                for k in range(mapping.shape[1]):
+                    # Loop over all rows/units again
+                    print("\t\tThird Loop: {}".format(k))
+                    for l in range(mapping.shape[0]):
+                        if mapping[l][k] > 0:
+                            # Choose (l, k)th units as the support and obtain it's mean waveform
+                            print("\t\t\tFourth Loop {}\n".format(l))
+                            runit_support = virtual_unit_map.get_realunit(k, l, data)
+                            support_wave = data.get_data("average", runit_support)
+                            support_rp = data.get_data("rate profile", runit_support)
+                            #
+                            # Calculate and store the Euclidean distance to
+                            # support from actor (multiplied by a distance
+                            # factor)
+                            #
+                            distances_waves[l][k] = np.linalg.norm(
+                                np.subtract(actor_wave, support_wave))  # * np.exp(abs(k-i))
+                            distances_rp[l][k] = np.linalg.norm(np.subtract(actor_rp, support_rp))
+                #                            if mapping[l][k] > 0:
+                #                                history[l][k] = distances[l][k]
+
+                # print("Distances:\n {}\n".format(distances))
+
+                #
+                # Extract the dataset corresponding to all distance measures
+                # with non-zero mapping values and calculate the reject threshold
+                #
+
+                distances = distances_waves  # + distances_rp
+
+                threshold_dataset = distances[mapping > 0]
+                dataset_reject_threshold = np.mean(threshold_dataset)
+                # dataset_std = np.std(threshold_dataset)
+                # threshold_range_L = threshold_mean + 0.5 * threshold_std
+                # dataset_reject_threshold = dataset_mean + 0.5 * dataset_std
+
+                # print("Threshold: {}\n".format(dataset_reject_threshold))
+
+                #
+                # The following loop is the main part of the mapping algorithm,
+                # which uses the distances between units to rearrange (swap) them.
+                #
+
+                # Generate loop range for the loop
+                if mapping.shape[1] < 10:
+                    loop_range = range(mapping.shape[1])
+                else:
+                    loop_range = [x for x in range(i + 9) if (i + x) <= mapping.shape[1]]
+
+                # Start looping over sessions/columns
+                for k in loop_range:
+                    # Find the argument of the minimum in the kth column
+                    min_arg = np.argmin(distances[:, k])
+                    print("i: {}, j: {}, k: {}, min_arg: {}\n".format(i, j, k, min_arg))
+                    if min_arg == j:
+                        # Do nothing if the minimum argument is the same as the row we're looping over
+                        pass
+                    elif distances[min_arg][k] <= dataset_reject_threshold:
+                        # If the distance between the two waveforms falls within
+                        # the threshold, we swap the two positions in the kth col
+                        if np.isnan(history[min_arg][k]):
+                            # If the unit has not been swapped earlier.
+                            history[j][k] = distances[min_arg][k]
+                            virtual_unit_map.swap(k, min_arg, j)
+                            # mapping = backup_mapping.copy()
+                            swaps += 1
+                        else:
+                            # If the unit has been swapped earlier, swap only if
+                            # the new distance is less than the old distance
+                            prev_dist = history[min_arg][k]
+                            curr_dist = distances[min_arg][k]
+                            if curr_dist < prev_dist:
+                                history[j][k] = distances[min_arg][k]
+                                virtual_unit_map.swap(k, min_arg, j)
+                                # mapping = backup_mapping.copy()
+                                swaps += 1
+                    elif distances[min_arg][k] > dataset_reject_threshold:
+                        # If the distance between the two waveforms is greater than
+                        # the threshold, move the unit to the first available free
+                        # plot
+                        loc = 0
+                        first_zero = np.where(mapping[:, k] == 0)[0][loc]
+
+                        try:
+                            while np.sum(mapping[first_zero]) > 0.5:
+                                loc += 1
+                                first_zero = np.where(mapping[:, k] == 0)[0][loc]
+                        except IndexError:
+                            first_zero = 0
+                        virtual_unit_map.swap(k, first_zero, min_arg)
+                        history[min_arg][k] = np.nan
+                        mapping = backup_mapping.copy()
+                        swaps += 1
+                        # print("Swapped {} with empty plot {} on day {}".format(min_arg, first_zero, k))
+                        # print("Mapping: {}".format(mapping))
+                    else:
+                        # If none of the above cases yield true, something is wrong
+                        print("Exception reached")
+                        print(distances_waves)
+                        print(distances_rp)
+                        print(distances)
+                        mapping = backup_mapping.copy()
+    storage.change_map()
+    print("Total swaps: {}\n".format(swaps))
+
+
+"""
+def calculate_mapping(self, data, base):
+
+    Calculates a mapping for the units based on features like distance.
+
+    The units will be compared pare-wise and sequential.
+
+    **Arguments**
+
+        *data* (:class:`src.neodata.NeoData`):
+            This object is needed to get the data 
+            which will be used to compare the units.
+
+
+    print("Map being calculated")
+    #dis = {}
+    #for each block except the last one
+    for i in range(len(data.blocks) - 1):
+        swapped1 = []
+        swapped2 = []
+
+        #do it so often that each real unit can find a partner
+        for n in range(data.total_units_per_block[i]):
+            distances = []
+
+            #for each unit in block i
+            for j in range(self.maximum_units):
+                if self.mapping[i][j] != 0:
+                    unit1 = self.get_realunit(i, j, data)
+                    y1 = data.get_data("average", unit1)[0]
+
+                    #for each unit in block i+1
+                    for k in range(self.maximum_units):
+                        if self.mapping[i+1][k] != 0:
+                            unit2 = self.get_realunit(i+1, k, data)
+                            y2 = data.get_data("average", unit2)[0]
+                            #calculates the distance between the average waveforms
+                            distance = np.linalg.norm(np.subtract(y2, y1))
+
+                            #calculate cross correlation {
+                            #calculates the inter-spike-interval
+                            isi1 = data.get_data("units", unit1)[0]
+                            isi2 = data.get_data("units", unit2)[0]
+                            len1 = len(isi1)
+                            len2 = len(isi2)
+                            if len1 > len2:
+                                isi1 = isi1[:len2]
+                            else:
+                                isi2 = isi2[:len1]
+                            cor1 = np.corrcoef(y1, y2)
+                            cor2 = np.corrcoef(isi1, isi2)
+                            c = ((cor1 + cor2) / 2.0)[1][0]
+                            #print("unit {} vs. unit {} = {}, crosscorr = {}".format(j, k, distance, c))
+                            # }
+
+                            #calculate norm {
+                            y = y1 + y2
+                            isi = isi1 + isi2
+                            av1 = np.mean(y)
+                            std1 = np.std(y)
+                            av2 = np.mean(isi)
+                            std2 = np.std(isi)
+                            yn1 = (y1 - av1) / std1
+                            yn2 = (y2 - av1) / std1
+                            isin1 = (isi1 - av2) / std2
+                            isin2 = (isi2 - av2) / std2
+                            #fig = plt.figure("dist: {}; corr: {}".format(distance, c))
+                            #fig.add_subplot(111)
+                            #plt.plot(yn1)
+                            #plt.plot(yn2)
+                            #plt.plot(isin1)
+                            #plt.plot(isin2)
+                            #plt.show()
+                            #print("unit {} vs. unit {} = {}, y = {} {}, isi = {} {}".format(j, k, distance,
+                            #                                                                yn1[:2], yn2[:2],
+                            #                                                                isin1[:2], isin2[:2]))
+                            #  }
+
+                            if j not in swapped1 and k not in swapped2:
+                                distances.append((j, k, distance))
+                                #dis[c] = distance
+            if distances:
+                #finds the min of the distances
+                mini = min(distances, key=lambda x: x[2])
+                #swap only if distance is under a threshold
+                if mini[2] < 200:
+                    #swaps the units to have them in the same unit row                
+                    self.swap(i+1, mini[0], mini[1])
+                    swapped1.append(mini[0])
+                    swapped2.append(mini[0])
+                #otherwise search for an empty row and swap there
+                #but only if the units are in the same row 
+                else:
+                    if mini[0] == mini[1]:
+                        for u, v in enumerate(self.mapping[i+1]):
+                            testing = True
+                            for l in range(i+2):
+                                if self.mapping[l][u] != 0:
+                                    testing = False
+                            if testing:
+                                self.swap(i+1, mini[1], u)
+                                swapped1.append(mini[0])
+                                swapped2.append(u)
+
+    #plt.plot(dis.keys(), dis.values(), "o")
+    #plt.show()
+"""
